Fix #10: Crash on headless servers: clipboard X11 connection failure
The clipboard initialization on TableView::default was calling `unwrap`
on the value returned by `Clipboard::new`.

On enviroments without a window manager session (e.g. A server accessed
via ssh), this would cause the application to panic.

Instead, make the `clipboard` field on TableView and Optional, and
simply ignore yank actions if no clipboard was available.