Make the clipboard feature opt-out
The arboard dependency pulls a lot of crates to give support for
multiple desktop enviroments.
This patch puts it behind a "clipboard" feature, which is enabled by
default.

Wayland support is also behind the "wayland" feature, also enabled by
default.

The idea is to give more flexibility to the people compiling this app
for their personal use.

For example: If I'm building JDbrowser for my home server, which I
access via ssh, I don't need clipboard support, since I won't be able
to use it.
And even if I build it for my laptop, I won't need wayland support if
I'm on an X11 server.

Since the features are enabled by default, nothing breaks. But if
someone wants to opt-out of a bunch of crates they won't use, let them!